+++ activity/docs.go
@@ -1,2 +1,0 @@
-// Package activity provides an enumeration of forager activities.
-package activity