+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [[unpublished]](https://github.com/mlange-42/beecs/compare/v0.1.0...main)
 
+### Features
+
+- Implements seasonal and "scripted" patch dynamics (#49)
+
 ### Bugfixes
 
 - Avoid drawing random parameters for experiments multiple times (#46)